update client structure
