+++ .gitignore
@@ -1,0 +1,4 @@
+.ipynb_checkpoints/*
+*.ipynb
+*.txt
+*.pyc
